refactor: 테스트 페이지 타입, 유틸함수 분리
